requirements.txt
crewai
crewai-tools
apscheduler
beautifulsoup4
requests
flask
python-dotenv
groq
